fix(app): remove props causing unnecessary re-renders
Removes caption styles and the audio captions live container from being
passed as props to the app component, as they were causing unnecessary
re-renders. They are now relocated inside the dedicated audio live
captions component.
